service and resource completed, DtoTest

+++ src/main/java/com/example/jakarta_books/repository/BookRepository.java
@@ -3,6 +3,7 @@ package com.example.jakarta_books.repository;
 import com.example.jakarta_books.entity.Book;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
 
@@ -15,29 +16,34 @@ public class BookRepository {
     EntityManager entityManager;
 
     @Transactional
-    public Book createBook(Book book) {
+    public Book createNew(Book book) {
         entityManager.persist(book);
         return book;
     }
 
-    public List<Book> readAllBooks() {
+    public List<Book> findAll() {
         return entityManager
                 .createQuery("SELECT b FROM Book b", Book.class)
                 .getResultList();
     }
 
-    public Book findBookById(long id) {
+    public Book findById(long id) {
         return entityManager.find(Book.class, id);
     }
 
-    @Transactional
-    public Book updateBook(Book updatedBook) {
-        return entityManager.merge(updatedBook);
-    }
 
     @Transactional
     public void deleteBook(long id) {
-        var book = findBookById(id);
+        var book = findById(id);
+        if (book == null) {
+            throw new EntityNotFoundException();
+        }
         entityManager.remove(book);
     }
+
+    @Transactional
+    public Book updateBook(Book book) {
+        return entityManager.merge(book);
+    }
+
 }
